requirements.txt
selenium==4.29.0
webdriver-manager==4.0.2
Pillow==11.1.0
numpy==1.21.0
scikit-learn==0.24.2
matplotlib==3.4.2
